server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
pydantic==2.9.2
python-multipart==0.0.9
numpy==1.26.4
scipy==1.13.1
opencv-python-headless==4.10.0.84
torch>=2.1
torchvision>=0.16
ffmpeg-python==0.2.0
orjson==3.10.7
tqdm==4.66.4

# Transformers backend
transformers>=4.51.0
accelerate>=0.30.0
pillow>=10.0.0
huggingface_hub>=0.23.0
onnxruntime==1.18.0

# MMPose backend (DW-Pose path)
mmcv==2.1.0
mmengine==0.10.7
mmdet==3.3.0
mmpose==1.3.0
xtcocotools==1.14.3
# You will also need DWPose configs + checkpoints from IDEA-Research/DWPose
